@@ -11,11 +11,20 @@
   },
   "devDependencies": {
     "axios": "^0.15.3",
+    "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "bootstrap-sass": "^3.3.7",
+    "font-awesome": "^4.7.0",
     "jquery": "^3.1.1",
+    "jquery-ui": "^1.12.1",
+    "jspath": "^0.4.0",
     "laravel-mix": "0.*",
     "lodash": "^4.17.4",
-    "vue": "^2.1.10"
+    "nprogress": "^0.2.0",
+    "photoswipe": "^4.1.2",
+    "vue": "^2.5.16",
+    "vue-router": "^3.0.1",
+    "vue-template-compiler": "^2.5.16",
+    "vuex": "^3.0.1"
   },
   "dependencies": {
     "swagger-ui-dist": "^3.8.0"
