@@ -6,8 +6,18 @@ use Illuminate\Http\Request;
 
 class TaxonController extends Controller
 {
+
     public function getTaxon(Request $request, $id)
     {
-        return view('taxon', ['id' => $id]);
+        $data = $this->getAuthInfo();
+        $data['id'] = $id;
+        return view('app', ['data' => $data]);
+    }
+
+    public function getKey(Request $request, $id)
+    {
+        $data = $this->getAuthInfo();
+        $data['id'] = $id;
+        return view('app', ['data' => $data]);
     }
 }
