@@ -12,7 +12,14 @@ let mix = require('laravel-mix');
  */
 
 mix.js('resources/assets/js/app.js', 'public/js')
-   .sass('resources/assets/sass/app.scss', 'public/css')
-   .mix.copy('node_modules/swagger-ui-dist/swagger-ui.css', 'public/css/swagger-ui/swagger-ui.css')
-   .mix.copy('node_modules/swagger-ui-dist/swagger-ui-bundle.js', 'public/js/swagger-ui/swagger-ui-bundle.js')
-   .mix.copy('node_modules/swagger-ui-dist/swagger-ui-standalone-preset.js', 'public/js/swagger-ui/swagger-ui-standalone-preset.js');
+    .sass('resources/assets/sass/app.scss', 'public/css')
+    .copy('node_modules/font-awesome/css/font-awesome.css', 'public/css/font-awesome.css')
+    .copy('node_modules/font-awesome/fonts', 'public/fonts')
+    .version().sourceMap();
+
+mix.copy('node_modules/swagger-ui-dist/swagger-ui.css',
+        'public/css/swagger-ui/swagger-ui.css')
+    .copy('node_modules/swagger-ui-dist/swagger-ui-bundle.js',
+        'public/js/swagger-ui/swagger-ui-bundle.js')
+    .copy('node_modules/swagger-ui-dist/swagger-ui-standalone-preset.js',
+        'public/js/swagger-ui/swagger-ui-standalone-preset.js');
